All other bin files
I know you explained really quickly to us how to upload to github properly, with sending it directly from VS and all, but I couldn't actually recall most of that, and saw it wasn't in the class recording like 30 minutes before the deadline (you may have noticed I'm both a little backlogged and a chronic procrastinator), and tried to look it up but all the methods I saw didn't seem to be options for me for some weird reason. So uh, enjoy. 
Apologies in advance ¯\_(ツ)_/¯